Validate appt date against sentencing date
This commit also changes most occurences of event number to a Long (changes on `AppointmentEventEntity` will be made in a separate commit)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/communitypaybackapi/client/CommunityPaybackAndDeliusClient.kt
@@ -97,7 +97,7 @@ interface CommunityPaybackAndDeliusClient {
   @GetExchange("/case/{crn}/event/{eventNumber}/appointments/schedule")
   fun getUnpaidWorkRequirement(
     @PathVariable crn: String,
-    @PathVariable eventNumber: Int,
+    @PathVariable eventNumber: Long,
   ): NDUnpaidWorkRequirement
 
   @Cacheable(CacheKey.Delius.GET_NON_WORKING_DAYS)
