+++ src/main/kotlin/uk/gov/justice/digital/hmpps/communitypaybackapi/service/scheduling/SchedulingService.kt
@@ -50,7 +50,7 @@ class SchedulingService(
 
   fun scheduleAppointments(
     crn: String,
-    eventNumber: Int,
+    eventNumber: Long,
     trigger: SchedulingTrigger,
     dryRun: Boolean,
   ): UUID {
